+++ src/python/pants/notes/2.18.x (and beyond).md
@@ -1,0 +1,2 @@
+(From 2.18.x onward, the per-commit changelog is no longer managed in repo, and is maintained through
+the individual releases: https://github.com/pantsbuild/pants/releases)
